# [rush] Improve cache read/write perf by attempting to use the "tar" binary.