--- a/common/changes/@microsoft/rush/ianc-faster-tar_2021-02-15-05-38.json
+++ b/common/changes/@microsoft/rush/ianc-faster-tar_2021-02-15-05-38.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "packageName": "@microsoft/rush",
+      "comment": "Improve cache read/write perf by attempting to use the \"tar\" binary.",
+      "type": "none"
+    }
+  ],
+  "packageName": "@microsoft/rush",
+  "email": "iclanton@users.noreply.github.com"
+}