… apps/rush-lib/src/utilities/TarExecutable.ts