## Summary

Currently, Rush uses the `tar` NPM package to create and extract cache entries. This works well, but isn't as fast as `tar` executable. This PR adds the ability to use `tar` if it is on the PATH.

## How it was tested

I've tested this on two repos: this one (rushstack) and a large internal repo. Things work fine, and archives created with older versions of Rush are able to be read by Rush with these changes, and vise-versa.

For the rushstack repo, restoring from cache took 8.23 seconds before, and now takes 3.05 seconds. In the large internal repo, restoring from cache took 10 minutes and 51.1 seconds, and now takes 1 minute and 35 seconds.